Fix the changes mentioned in cr
