--- github-connector/README.md
@@ -1,11 +1,0 @@
-# heroku-go-test
-Heroku GitHub go test
-
-* hosting on Heroku
-* checking GitHub interaction
-* in go
-
-# Deployment (in progress)
-
-```heroku create```
-```git push heroku master```
